event_service/requirements.txt
flask
flask-restful
requests
psycopg2-binary
